refactor: 메서드 및 파라미터 이름 수정 #17
